chore/#119: AlertButtonConfig 파일 분리
