+++ Meomun/Meomun/Presentation/Common/Modifier/CustomAlertModifier.swift
@@ -7,24 +7,6 @@
 
 import SwiftUI
 
-struct AlertButtonConfig {
-    enum Role {
-        case normal
-        case cancel
-        case destructive
-    }
-
-    let title: String
-    let role: Role
-    let action: (() -> Void)?
-
-    static let confirm = AlertButtonConfig(
-        title: "확인",
-        role: .cancel,
-        action: nil
-    )
-}
-
 struct CustomAlertModifier<AlertModel>: ViewModifier {
     @Binding var alert: AlertModel?
